+++ rust/src/lib.rs
@@ -68,7 +68,6 @@ pub mod smb;
 pub mod krb;
 pub mod dcerpc;
 
-pub mod ikev2;
 pub mod snmp;
 
 pub mod ntp;
