+++ src/app-layer-parser.c
@@ -61,7 +61,6 @@
 #include "app-layer-nfs-udp.h"
 #include "app-layer-ntp.h"
 #include "app-layer-tftp.h"
-#include "app-layer-ikev2.h"
 #include "app-layer-krb5.h"
 #include "app-layer-snmp.h"
 #include "app-layer-sip.h"
@@ -1559,7 +1558,6 @@ void AppLayerParserRegisterProtocolParsers(void)
     RegisterNFSUDPParsers();
     RegisterNTPParsers();
     RegisterTFTPParsers();
-    RegisterIKEV2Parsers();
     RegisterKRB5Parsers();
     rs_dhcp_register_parser();
     RegisterSNMPParsers();
