+++ src/rust-context.h
@@ -20,7 +20,6 @@
 
 #include "detect-engine-state.h" //DetectEngineState
 #include "app-layer-krb5.h" //KRB5State, KRB5Transaction
-#include "app-layer-ikev2.h" //IKEV2State, IKEV2Transaction
 #include "app-layer-ikev1.h" //IKEV1State, IKEV1Transaction
 #include "app-layer-ntp.h" //NTPState, NTPTransaction
 #include "app-layer-snmp.h" //SNMPState, SNMPTransaction
